Move `getUrlForDataEntry` to feature
