+++ frontend/src/features/data_entry_choice/utils/util.ts
@@ -57,3 +57,12 @@ export function getPollingStationWithStatusList({
     return result;
   });
 }
+
+export function getUrlForDataEntry(
+  electionId: number,
+  pollingStationId: number,
+  pollingStationStatus?: DataEntryStatusName,
+): string {
+  const entryNumber = pollingStationStatus?.startsWith("second_entry") ? 2 : 1;
+  return `/elections/${electionId}/data-entry/${pollingStationId}/${entryNumber}`;
+}
